+++ experimental/packages/opentelemetry-sdk-node/test/start.test.ts
@@ -36,8 +36,6 @@ import {
 } from '@opentelemetry/sdk-logs';
 import type {
   ConfigFactory,
-  ConfigurationModel,
-  HttpTlsConfigModel,
   LogRecordExporterConfigModel,
 } from '@opentelemetry/configuration';
 import { createConfigFactory } from '@opentelemetry/configuration';
@@ -57,13 +55,7 @@ import {
   ATTR_SERVICE_INSTANCE_ID,
 } from '../src/semconv';
 import { ATTR_OS_TYPE } from '@opentelemetry/resources/src/semconv';
-import {
-  getAggregationType,
-  getHttpAgentOptionsFromTls,
-  getLogRecordExporter,
-  getSpanLimitsFromConfiguration,
-  setupContextManager,
-} from '../src/utils';
+import { getLogRecordExporter, setupContextManager } from '../src/utils';
 import {
   ConsoleMetricExporter,
   MeterProvider,
@@ -72,7 +64,6 @@ import {
 import type {
   SpanProcessor,
   NodeTracerProvider,
-  SpanLimits,
 } from '@opentelemetry/sdk-trace-node';
 import { BasicTracerProvider } from '@opentelemetry/sdk-trace-base';
 import {
@@ -972,131 +963,11 @@ describe('startNodeSDK', function () {
       assert.equal(getLogRecordExporter(exporter), undefined);
     });
 
-    it('', async () => {
+    it('null context manager', async () => {
       setupContextManager(null);
       assert.equal(
         context['_getContextManager']().constructor.name,
         'NoopContextManager'
-      );
-    });
-
-    it('return undefined with no config for tracer limits', async () => {
-      assert.equal(
-        getSpanLimitsFromConfiguration({} as ConfigurationModel),
-        undefined
-      );
-    });
-
-    it('return span limits', async () => {
-      const config: ConfigurationModel = {
-        tracer_provider: {
-          processors: [],
-          limits: {
-            attribute_count_limit: 10,
-            event_count_limit: 20,
-            link_count_limit: 30,
-            attribute_value_length_limit: 40,
-            event_attribute_count_limit: 50,
-            link_attribute_count_limit: 60,
-          },
-        },
-      } as ConfigurationModel;
-      const expectedSpanLimits: SpanLimits = {
-        attributeCountLimit: 10,
-        eventCountLimit: 20,
-        linkCountLimit: 30,
-        attributeValueLengthLimit: 40,
-        attributePerEventCountLimit: 50,
-        attributePerLinkCountLimit: 60,
-      };
-
-      const spanLimits = getSpanLimitsFromConfiguration(config);
-      assert.deepEqual(spanLimits, expectedSpanLimits);
-    });
-
-    it('return undefined for no aggregation type', async () => {
-      assert.equal(getAggregationType({}), undefined);
-    });
-  });
-
-  describe('getHttpAgentOptionsFromTls', function () {
-    it('should return undefined if no TLS config is provided', async () => {
-      assert.equal(getHttpAgentOptionsFromTls({}), undefined);
-    });
-
-    it('should return https agent options if TLS config is provided', async () => {
-      const tlsConfig: HttpTlsConfigModel = {
-        ca_file: 'test/fixtures/ca.pem',
-        key_file: 'test/fixtures/ca-key.pem',
-        cert_file: 'test/fixtures/cert.pem',
-      };
-      const agentOptions = getHttpAgentOptionsFromTls(tlsConfig);
-      assert.ok(agentOptions);
-      assert.notEqual(agentOptions.ca, undefined);
-      assert.notEqual(agentOptions.key, undefined);
-      assert.notEqual(agentOptions.cert, undefined);
-    });
-
-    it('show warning messages for invalid ca file', async () => {
-      const stubLoggerWarn: Sinon.SinonStub = Sinon.stub(diag, 'warn');
-      const tlsConfig: HttpTlsConfigModel = {
-        ca_file: 'invalid-ca.pem',
-        key_file: 'test/fixtures/ca-key.pem',
-        cert_file: 'test/fixtures/cert.pem',
-      };
-      const agentOptions = getHttpAgentOptionsFromTls(tlsConfig);
-      assert.ok(agentOptions);
-      assert.equal(agentOptions.ca, undefined);
-      assert.notEqual(agentOptions.key, undefined);
-      assert.notEqual(agentOptions.cert, undefined);
-
-      assert.equal(
-        stubLoggerWarn.args[0][0].startsWith(
-          'Failed to read TLS CA file at invalid-ca.pem: Error: ENOENT: no such file or directory, open '
-        ),
-        true
-      );
-    });
-
-    it('show warning messages for invalid ca-key file', async () => {
-      const stubLoggerWarn: Sinon.SinonStub = Sinon.stub(diag, 'warn');
-      const tlsConfig: HttpTlsConfigModel = {
-        ca_file: 'test/fixtures/ca.pem',
-        key_file: 'invalid-ca-key.pem',
-        cert_file: 'test/fixtures/cert.pem',
-      };
-      const agentOptions = getHttpAgentOptionsFromTls(tlsConfig);
-      assert.ok(agentOptions);
-      assert.notEqual(agentOptions.ca, undefined);
-      assert.equal(agentOptions.key, undefined);
-      assert.notEqual(agentOptions.cert, undefined);
-
-      assert.equal(
-        stubLoggerWarn.args[0][0].startsWith(
-          'Failed to read TLS key file at invalid-ca-key.pem: Error: ENOENT: no such file or directory, open '
-        ),
-        true
-      );
-    });
-
-    it('show warning messages for invalid cert file', async () => {
-      const stubLoggerWarn: Sinon.SinonStub = Sinon.stub(diag, 'warn');
-      const tlsConfig: HttpTlsConfigModel = {
-        ca_file: 'test/fixtures/ca.pem',
-        key_file: 'test/fixtures/ca-key.pem',
-        cert_file: 'invalid-cert.pem',
-      };
-      const agentOptions = getHttpAgentOptionsFromTls(tlsConfig);
-      assert.ok(agentOptions);
-      assert.notEqual(agentOptions.ca, undefined);
-      assert.notEqual(agentOptions.key, undefined);
-      assert.equal(agentOptions.cert, undefined);
-
-      assert.equal(
-        stubLoggerWarn.args[0][0].startsWith(
-          'Failed to read TLS cert file at invalid-cert.pem: Error: ENOENT: no such file or directory, open '
-        ),
-        true
       );
     });
   });
